\name{testClass}
\docType{data}
\alias{testClass}
\title{Sample Test Data for the Iterative BMA Algorithm}
\description{
This is an adapted leukemia (ALL, AML) dataset from Golub et al. (1999).
This is a zero-one vector for 34 test samples.  Class 0 represents
an ALL sample, while class 1 represents an AML sample.
}
\usage{data(testClass)}
\format{The vector is called \code{testClass}. 
}
\details{
} 
\source{
The \code{golubEsets} bioconductor data package, or
\url{http://www.broad.mit.edu/cgi-bin/cancer/datasets.cgi}.
}
\references{
Golub, T.R., Slonim, D.K., Tamayo, P., Huard, C., Gaasenbeek, M., Mesirov, J.P., Coller, H., Loh, M.L., Downing, J.R., Caligiuri, M.A., et al. (1999) 
Molecular classification of cancer: class discovery and class prediction by gene expression monitoring. Science 286: 531-7.
}
\keyword{datasets}

